Code executing in Forth Virtual Machine
The Forth Virtual Machine is currently not fully implemented nor tested,
but it is currently executing a test program that prints out "Hi!"
followed by a line feed, and then exits. This tests a few Forth
primitives and the mechanism to execute primitives, but not yet calls
and some other important things that need to be present in the base
system.