Added bitwise operators
Some of the bitwise operators have been added to this Forth and they
appear to work, they are incredibly slow whilst they are being used
however and will have to be optimized, it might be worth rewriting them
directly in assembly.

Some other words which build on top of these have been added, as well
as some commented out code.