Reimplementing operators/new assembly words
Many of the basic primitives are not quite working, there are subtle
bugs in many of the functions, so the plan is to reimplement many of the
operators using a different scheme that is also hopefully more
efficient. Some functions have already been made to be more efficient,
such as the Least Significant Bit extraction, and multiplication has
been turned into an assembly routine.