Hack: extend SUBLEQ VM to work around defective ops
I have extended the SUBLEQ VM with four new instructions to work around
defective operations, like "Get The Lowest Bit And Turn It Into A Forth
Boolean" ("lsb"), or "Right Shift by One", there may be a few other
operations that are incorrect (most likely comparison operators).

Adding these memory mapped instructions defeats the purpose of having a
SUBLEQ One Instruction Set Computer machine, so they will be removed
later.

The idea is to get as much of the interpreter working as possible, a
complete working Forth with the higher level Forth code debugged and
fully working, one capable of introspection, and then to replace the new
instructions one by one, until we are back to the SUBLEQ machine.

Currently, the interpreter loop seems to partially work, the throw/catch
mechanism causes problems, as does numeric input and out, but words
can be looked up and executed, the interpreter loop seems to be
particularly brittle at the moment however, and the lack of proper
terminal control causes problems (the Linux and Windows terminal by
default echo characters typed in, this Forth interpreter assumes that
we are communicating over a UART and this does not happen).

In short, there are still major problems that need to be resolved, but
some progress can still be made. It appears that the instruction
synthesized "msb" (or "Get The Most Significant Byte") appears to be
wildly incorrect, and the "2*" word was largely a red-herring.

The changes are committed to a branch called "hack", because this is
what this commit contains, until the issues are resolved it should not
be merged back into the mainline (whenever that is) and should be kept
separate.