Hack: Fixed um/mod, negate, numeric I/O
A problem with the implementation of "negate" has been fixed. I believe
this was cause by messing around with random things when a lot of other
things were broken, and not being able to rely on anything. The break
was not noticed because everything else was wrong.

Defining new words within the running interpreter still does not work
correctly, and there is bound to be other errors that will need
addressing before removing the various hacks put in place so I could get
other things working.