Hack: compiling works, removed two hacks
A number of minor bugs with major results have been fixed, there are a
few worrying problems that remain, but the interpreter is in a much
better shape than before.

* A bug in the string implementation has been fixed, which was caused by
using a Forth boolean instead of a 1/0 boolean value.
* New words can be compiled at run-time, this includes strings and
control structures.
* The test code has been removed.
* Lots of code has been renamed and reworked.
* A few serious bugs remain, they are "u<" not working for all values,
"throw"/"catch" seem to return the wrong values, and an odd one in the
implementation of ",", which has been fixed by doing an equivalent
operation, why the original implementation did not work I do not know.
* There have been a few minor optimizations, which should be continued
once the interpreter is fully working.

All in all, this is a good commit!