Hack: checksum, "info" word, found 8000 problem
* A checksum has been added, it is a simple additive checksum over all
the bytes modulo 2^16, I could improve the algorithm, but probably will
not.
* An "info" word has been added which has details about the project.
* The problem has been found, in fact I am surprised the interpreter
works as well as it does with this basic problem, there are various
words used throughout the interpreter that fail to work correctly when
fed the annoying number "8000", which is an annoying number in twos
compliment, they include "if", "0<", and some other functionality.
Fixing these issues (in "if" mainly") will allow this interpreter to
remove the hacks that have been added.