v1.0
This commit brings the system into a fully workable state, there are
still known issues, but nothing dreadful. The VM has been simplified,
the image "subleq.dec" committed to the repo in case gforth is not
installed, and a few things have been rearranged, but nothing special.

The "readme.md" file has been rewritten and is now more terse.