Added "create", ":noame", "variable", "recurse"
* There have been some minor optimizations made, mainly to remove use of
bitwise operators where possible.
* A comment has been made about merging the word definitions which are
just contain a single assembly call usage. This should save on space and
make quicker run time compiled code. It should also save on some space.
* Some new words have been added, "create", ":noname", "variable", and
"recurse", all of which would be difficult to implement at run time, but
not impossible. The major missing definitions are the vocabulary
word-set (still), "constant" and "does>", "does>" will be the most
difficult to implement and will probably require modifications elsewhere
as well.
* The "GET" assembly directive has been removed, as it is possible to
use "[@]" to do the same thing, it is not yet possible to get rid of
"PUT" with "[!]" (yet) because it must load, causing a input operation
to be triggered.
* "(drop)" has been removed as a word, it is not needed.