Sped up interpreter with 2/ reimplementation
* "does>" almost works, it is commented out still.
* Added execution time to the test framework.
* The interpreter has been sped up greatly with the reimplementation of
both "2/" and "rshift" in assembly.
* There is now an option to quit the interpreter on the receipt of an
EOF character, this would be used to indicate a character has not
arrived on a system with non-blocking I/O, but is used to indicate
End-Of-Input on a system with blocking I/O, which we currently have.
* Updated comments, one comment contains a tiny interpreter for the VM
which will be used once everything is sorted out (it is a little slower
than the current one, but smaller!).
* Added "postpone" as a word
* Removed "also" in the meta-compiler, it is not needed.
* I have not yet come up with a way to fix "<" and ">".