Added meta-compiler steps
A temporary file has been added to deal with meta-compilation. Although
it currently does not work, the meta-compiled file generated in
"meta.dec" is very similar to "subleq.dec". There do seem to be some
subtle problems with the meta-compiler not present with gforth which
must be down to differences in the system.