Meta-compilation of an image mostly works
Meta-compilation of an image using the SUBLEQ system itself now
(mostly) works! Which is great. There are a few words which do not work
and the image which is generated runs a Forth but cannot yet be used
to generate another image, which will need to be resolved. Also, dumping
the image numbers is slow, and might benefit from custom divide/mod
routines.

The main problem holding up meta-compilation was the difference in
behavior between gforth's "'" and SUBLEQ eForths version of "'".