Image generation is identical
Apart from an errant new line, image generation is now identical to the
gforth compiled image. The next step will be in attempting to speed up
saving the image. Before the gforth toolchain is deleted there are some
serious bugs that need to be sorted that a fast turn around would help
in debugging.