Cleaned up vocabulary a lot
* Added the "marker" keyword, which will work so long as no vocabulary
manipulation is performed between the creation of the marker and the use
of the marker. Invalid use could be detected, but will not be.
* A word called "many" has been added, which is a cute little word that
causes the interpreter to continually evaluated the last typed line.
* The default vocabulary has been cleaned up a lot, this has been done
by adding a "system" vocabulary and putting most words that do not
belong in a default forth system in it. They are still accessible
though, which is good.