Semi-working commit; changed I/O, editor words
This commit has problems that the previous one does not, the "test"
makefile target is not working correctly, although the meta-compilation
does settle down into a stable image that can compile itself. The image
is also different to the one produced by gforth compilation process.

It appears that the "[else]" keyword no longer works correctly as well.

It might have something to do with the new I/O mechanism, specifically
the input buffer, which has been made to be more standards compliant
so "evaluate" can work, which is needed for "load", which is needed for
the Block Editor. The Block Editor is mostly working, it is missing the
insert command, and, obviously, the load command does not work
correctly as stated.