Editor insertion now works, still buggy
This version still does not metacompile properly, however the editor
commands have been fixed, it is now possible to insert text and use it
as an editor. The editor will still have to be improved, there is no
error detection and when an invalid command is given it exits back to
the main Forth program.

The bugs will have to be fixed before this is merged back into v1.6,
which will then need more development before it is merged back into
master.