Found possible cause of bug, not fixed
There is an odd bug, most likely due to alignment issues, or something
odd happening with the vocabularies, which caused "nop" to display an
inconsistent value that did not make sense, a hack has been put in place
the "fixes" the issue (commented in the code) but the root cause will
have to be found and fixed. It is mostly likely a problem in the
meta-compiler. In this commit however, "evaluate" does work.