Evaluate bug found, not fixed (gforth target only)
The bug with evaluate only manifests when using the gforth target, there
are two problems, one is that the gforth target differs from the one
produced by SUBLEQ in minor ways (namely when it comes to the numeric
representation of some 16-bit signed values (65535 in gforth generation
image, -1 in SUBLEQ eForth generated image, the SUBLEQ one being
correct), the second is more serious in that whenever "t' nop lit" is
used, which should compile the execution token of "nop" to be pushed to
the stack at run time into a dictionary word defined in the cross
compiled target dictionary the gforth version writes junk into that
value, random junk that changes with each compilation of the file, this
might be to do with some value not being initialized correctly on the
gforth version.

Other changes:
* Changed some notes and comments about the interpreter
* Remove unneeded '$' on hex value (all compilation is done in hex)
* Simplified the 'list' command.
* Added a test file 'tst.fth', which will be removed later, it can be
used to test the block editor.