Fixed gforth nop bug; images now the same
The images generated by GForth and the SUBLEQ meta-compiler are now the
same, the problem was the wrong "nop" was being accessed. eForth is case
sensitive, so "nop" was the right one to compile as an execution token
into the dictionary in evaluate with "t' nop lit", however GForth is
case insensitive, there is a "NOP" word that is part of the
meta-compiler word-set used in the assembler, this was the one which was
getting compiled into the dictionary in GForth. This fixes the bug with
gforth compiled version.

Other minor changes include:
* Added a "task.fth" file for testing task words, which are currently
only partially implemented/working.
* The "clean" makefile target now uses git to do the cleaning.
* "blk" and "scr" are no longer user variables.
* "spaces" has been replaced by the more generic "banner" word.
* Some code has been moved around.
* The stack should have had zero values on it when the cross compilation
was complete, this was not the case, the source of those variables has
been found and fixed.
* There are now depth checks on more words, however the depth check is
still not the best.
* The "logical" word has been removed.