Hidden more words, formatted to 64 chars wide
More words have been hidden that are non-standard ones, but can still be
accessed via the "system" word-set. There have been some other minor
changes as well, that slightly shorten the size of the system.

The source code in "subleq.fth" has been reformatted so it is only 64
characters wide on each line, to allow it to potentially fit into a
series of Forth blocks.