Added a document generation script
Whilst there is still a massive amount to do in terms of the
documentation, a document generation script along with some makefile
rules has been added that can turn the "subleq.fth" script into a
markdown file, which can then be converted into other formats. It might
be better if the conversion script was written in Forth instead of Perl,
but it will do for now. It only does a little, there is still a lot that
would need to be done in terms of formatting, both in terms of the input
and formatting directives (like embedding more CSS in the output).