Got full self interpreter working
The full self interpreter capable of doing I/O is working, it could be
optimized. It would be the basis of turning the self-interpreter into
one that could turn N where N is > 16 into a simulator for an N = bit
bit length SUBLEQ interpreter. This would require a strategically placed
"AND" operation to be implemented, which can be done with
modulo/remainder.