Improved self-interpreter, rewrite "u."
There are many miscellaneous changes in this commit, to-do comments,
more elaboration in "subleq.fth", an improved "u." that uses "u.r", some
test code added into "ingest.fth", which does the opposite of "dump",
and the self-interpreter is a little improved. There are still bugs in
the self-interpreter (or it might be because of the extra space it takes
up).