Using mux, other minor optimizations
Some size (and speed) optimizations have been done, shaving instructions
off here and there, however the biggest change is the use of "mux"
instead of implementing the bitwise operators in assembly, that section
will need rewriting. Using "mux" to implement bitwise "and", "or", and
"xor" does not seem to slow anything down.

There is a curious, minor bug, the IOCCC candidate bit length detection
routine does not operate correctly for bit lengths less than 16, it
works on the N-bit machine and on the normal "subleq.c" program. The
self interpreter works fine with "prog.c".