Div/mod, better debug
The debug command continues to be improved, the recompiler has been
added in and ported so it is an N-bit one (there is more to the
integration), more commands have been added, there's still a lot to do.

Assembly code for div/mod has been added, the tracing facility of the
debug command helped in making it.