Bug in compile-only, smaller image, minor debug improvements
* There is a minor bug in the meta-compilers version of "compile-only",
and presumably "immediate" as well, in the fact that only the words in
the main vocabulary that is being meta-compiled are affected, and not in
the others ones, for example:

   : x ... ;
   :s y .... ;s compile-only

Will apply "compile-only" to "x" and not "y"! This can be worked around.

* The image is slightly smaller, more words have been refactored and
added but the image is still smaller.
* There have been minor improvements to "debug", and the self
interpreter has been added.
* "w" has been replaced with "r0" in the SUBLEQ assembly in
"subleq.fth".
* A new makefile target that strips out comments in the Forth code
before compiling greatly speeds up testing. Those comments really slow
things down when doing the self-hosting system tests.

The assembler built into "debug.c" still needs writing.