Added example code, rewrite using meta-vars
* A lot of example code has been added, it will be removed later,
but a lot of it should be ported to this SUBLEQ forth, it comes from
another project of mine containing a Forth interpreter based off of the
H2 VHDL CPU.
From:
  - <https://github.com/howerj/embed>
And:
  - <https://github.com/howerj/forth-cpu>

* The variables, constants and user variables defined as part of
meta-compilation are now (mostly) defined in a better way, using
meta-compiler versions of "variable", "constant" and "user". This means
defining variables and constants take up less room and is less awkward.
"user" still needs to be added in at run time.
* "(order)" is protected by a compile guard. This is in anticipation of
moving it to the main vocabulary, the word (along with "+order" and
"-order") is just too useful, as is "within".
* Lots of things have been moved around, so sections of the book might
not make sense and will have to be rewritten.