Improved assembler, removed "opPush"
* In the drive to make the SUBLEQ VM kernel smaller "opPush" has been
removed at the sacrifice of some speed. The resulting image is smaller
however, even with the extra words. This also means that several
paragraphs will need rewriting in the book, and no longer make sense due
to things being moved around or not existing anymore. The next
instruction to remove would be "pause", also compile time flags to
remove "opDivMod" and "rshift" would be useful.
* The assembler has been improved, but it is ugly, lots of confusing
control flow (I should have used a lexer/parser combination, but it
would be much bigger). There is still a few things that need fixing in
it, but it does produce code now. After it is fully working it will need
to be integrated into "debug.c" and "sqasm.cpp" will need removing.