Added test programs, fixed debug, minor Forth changes
Many test programs which will be removed have been added, such as
"bit.c" and "if.c", "if.c" is new and shows how the comparison
operators can be implemented, which will be needed to fix the bugs in
the Forth code comparison operators.

The re-compiler code for ISTORE has been fixed to reflect the smaller
ISTORE instruction.

The Forth code has some more explanations, and an unneeded "align"
removed.