Fixed comparison operator, updated float
There is a lot to do to get the floating point package into a good
state, more to get it to meta-compile as an optional component, however
many of the bugs have been ironed out. There is a TODO section with a
list of tasks, as there is in "subleq.fth", where the comparison
operator has been fixed, many sections need rewriting (and the allocator
improving) which is all marked out.