user variable bug, added f~, makefile
Two bugs have been fixed, on in the definition of "user", which now
works correctly (pre increment user counter and by cell size instead
of by one) and a minor one in the makefile where "subleq" was missing
from the "float" target.

The word "f~" has been added back in as well, it appears to work as it
should after some testing comparing it to the version in gforth.