debug.c cli, ioccc, floats/unit.fth
* The Command Line Interface to "debug.c" has been reworked a little,
removing options, as there were too many.
* The IOCCC entry has been updated, removing any numbers, and a bug has
been found - the self interpreter does not work correctly when prefixed
to the forth image, it works with the main "subleq.c" program however.
* The unit test framework has been reworked, it has not been tested and
will need fixing / porting to this forth.
* The floating point code, with some other stuff, has been appended to
"subleq.fth", it will need integrating.

I also need to work on the new version of the self-interpreter, but I am
putting that off for now...