More documentation, more space efficient opMux
The main change is that opMux is more space efficient, if an assembly
version of "else" were to be added several other assembly routines could
be made to be more efficient as well (both in terms of space and time).

A few more things have been document. And in the extension code more
version of the "str*" routines have been added.