12246: Fixed (basic) optimizer, made optional
An optimizer has been fixed, made optional, and turned off. The
optimization is quite simple, when unconditionally jumping to a
known destination in SUBLEQ assembler the previous instructions
jump can be modified in many cases to point to that location, even if
that previous instructions jump is not always executed. When it is
executed, one fewer jump is done, speeding things up slightly.

A basic optimization setting has been implemented but turned off by
default as it interferes with the recompiler in "debug.c" when it is
turned on.