11856: image optimizations, 2nd compressor
* A skeleton of the second generation of the compressor has been added,
it does not work at the moment, it will be similar to a more traditional
LZSS compression CODEC.
* Some words have been turned into constants to save space, and space
has been found elsewhere.
* The word "w", in "link.fth", which prints out a detailed table of
words along with meta-information about those words, has been extended
to (partially) detect words that are just call the built-in Forth VM
instructions.
* A TODO has been added to "subleq.fth" about rewriting "rp!" and "rp@"
into Forth and using them, along with sections of "pause", to save on
space.
* "r@" has been implemented in pure Forth, and the Forth VM instruction
for it removed.