11844: Removed unneeded INC in pause, compression
* Added an LZP compressor, it performs better than the 1c.c, and 2c.c
has not been made to work yet, but still not good enough.
* A set of unneeded INC instruction in the "pause" VM instruction have
been removed, which has saved space.
* The TODO comments in "subleq.fth" remain, and another one added, about
what can be done to save even more space. If done, some sections will
need to be rewritten. The book really needs printing out and editing
(and depositing with national British Library, which I should do, but
that's just an aside).