11696: Added working self-interpreter
A working "self-interpreter" has been added, that is a SUBLEQ program
that interprets and executes SUBLEQ programs. It will need modifying so
that it can present a 16-bit SUBLEQ machine on machine widths of greater
than 16-bits. There are many TODO comments that have been added relating
to this.

Another section of possible development has to do with the threading
model used for the built-in primitives, it might be more efficient to
take a more direct approach when it comes to them, placing the word
header for them next to the SUBLEQ assembly.