12224: Added "oneline.c", rearranged buffers
* Some buffers have been moved relating to the block interface and to
the optional allocator, which should make expanding the block buffers
easier.
* "oneline.c", a "one-liner" IOCCC candidate has been added, it needs to
get under 140 bytes to be worthy of submission, and obviously changed so
everything is on one line. The idea would be to make SUBLEQ interpreter
under 140 bytes capable of running the eForth interpreter. It would be
slightly easier if integers, or "int", could be used instead of "short",
which would require the n-bit self-interpreter to be working, which is
just mostly a bunch of comments at the moment.