12138: "extra/portable.dec", working self interp
The N-Bit self-interpreter is working! The system can detect that it
is running on a non-16-bit SUBLEQ VM and simulate one for the rest of
the Forth system. It is *slow*, so a warning should probably be printed
out, and larger (so it is compiled out normally).

The image for doing this is "extra/portable.dec", it adds about 2KiB of
program size and pushes the decimal file up to 30KiB.